python_learning/myflaskapp/requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest
webtest
factory-boy==2.4.1

# Management script
Flask-Script

# Debug toolbar
Flask-DebugToolbar==0.9.0